Implement Zod on the server for Edit and Checkout pet + create petIdSchema

+++ src/actions/actions.ts
@@ -1,13 +1,11 @@
 "use server";
 
 import prisma from "@/lib/db";
-import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
-import { petFormSchema } from "@/lib/validations";
-import { Pet } from "@prisma/client";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
 	await sleep(1000);
 
 	const validatedPet = petFormSchema.safeParse(pet);
@@ -30,15 +28,24 @@ export async function addPet(pet: PetEssentials) {
 	revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
 	await sleep(1000);
+
+	const validatedPetId = petIdSchema.safeParse(petId);
+	const validatedPet = petFormSchema.safeParse(newPetData);
+
+	if (!validatedPetId.success || !validatedPet.success) {
+		return {
+			message: "Invalid pet data",
+		};
+	}
 
 	try {
 		await prisma.pet.update({
 			where: {
-				id: petId,
+				id: validatedPetId.data,
 			},
-			data: newPetData,
+			data: validatedPet.data,
 		});
 	} catch (error) {
 		return {
@@ -48,13 +55,20 @@ export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
 	revalidatePath("/app", "layout");
 }
 
-export async function checkoutPet(petId: Pet["id"]) {
+export async function checkoutPet(petId: unknown) {
 	await sleep(1000);
+
+	const validatedPetId = petIdSchema.safeParse(petId);
+	if (!validatedPetId.success) {
+		return {
+			message: "Invalid pet data",
+		};
+	}
 
 	try {
 		await prisma.pet.delete({
 			where: {
-				id: petId,
+				id: validatedPetId.data,
 			},
 		});
 	} catch (error) {
